Todos los bugs arreglados. Seguridad mejorada. Ahora, se diferencia entre usuasior admin o estándar

+++ controlador.php
@@ -101,7 +101,7 @@
                 break;
 
                 case "2": //Usuarios estándar
-                    $this->gestionInstalaciones();
+                    $this->gestionReservas();
                 break;
 
                 case "3": //Usuarios deshabilitados
@@ -124,36 +124,52 @@
 
         public function gestionUsuarios($criterio = null,$mostrarBorrados = null) {
 
-            if ($criterio == null) {
-                $criterio = "id";
-            }
-            if ($mostrarBorrados == "on") {
-                $criterio2 = "";
+            if ($this->seguridad->esAdmin()) {
+                if ($criterio == null) {
+                    $criterio = "id";
+                }
+                if ($mostrarBorrados == "on") {
+                    $criterio2 = "";
+                } else {
+                    $criterio2 = "WHERE NOT borrado = 'si'";
+                }
+    
+                $data["usuarios"] = $this->usuario->getAll($criterio,$criterio2);
+                $data["borrados"] = $mostrarBorrados;
+                $this->vista->mostrar("usuario/gestionUsuarios", $data);
             } else {
-                $criterio2 = "WHERE NOT borrado = 'si'";
+                $this->gestionReservas();
             }
 
-            $data["usuarios"] = $this->usuario->getAll($criterio,$criterio2);
-            $data["borrados"] = $mostrarBorrados;
-            $this->vista->mostrar("usuario/gestionUsuarios", $data);
         }
 
         public function gestionInstalaciones($criterio = null) {
 
-            if ($criterio == null) {
-                $criterio = "id";
+            if ($this->seguridad->esAdmin()) {
+                if ($criterio == null) {
+                    $criterio = "id";
+                }
+    
+                $data["instalaciones"] = $this->instalacion->getAll($criterio);
+                $data["horarios"] = $this->horario->getAll(date('w'));
+                $this->vista->mostrar("instalacion/gestionInstalaciones", $data);
+            } else {
+                $this->gestionReservas();
             }
-
-            $data["instalaciones"] = $this->instalacion->getAll($criterio);
-            $data["horarios"] = $this->horario->getAll(date('w'));
-            $this->vista->mostrar("instalacion/gestionInstalaciones", $data);
 
         }
         
         public function gestionReservas() {
+
+            if ($this->seguridad->esAdmin()) {
+                $data["reservas"] = $this->reserva->getAll();
+            } else {
+                $data["reservas"] = $this->reserva->get($this->seguridad->get('id'));
+            }
             $data["instalaciones"] = $this->instalacion->getAll();
-            $data["reservas"] = $this->reserva->getAll();
+            $data["idUsuario"] = $this->seguridad->get('id');
             $this->vista->mostrar("reserva/gestionReservas", $data);
+            
         }
 
         // ==========================================================================================================
@@ -163,94 +179,127 @@
 
         public function crearUsuario() {
 
-            $id = $this->usuario->getLastId();
-            $usuario = $_REQUEST["usuario"];
-            $contrasenya = $_REQUEST["contrasenya"];
-            $email = $_REQUEST["email"];
-            $nombre = $_REQUEST["nombre"];
-            $apellido1 = $_REQUEST["apellido1"];
-            $apellido2 = $_REQUEST["apellido2"];
-            $dni = $_REQUEST["dni"];
-            $imagen = $_FILES["imagen"];
-            $borrado = 'no';
-            if (isset($_REQUEST["roles"])) {
-                $roles = $_REQUEST["roles"];
-            } else {
-                $roles = array("2");
-            }
 
-            if ($this->usuario->add($id,$usuario,$contrasenya,$email,$nombre,$apellido1,$apellido2,$dni,$imagen,$borrado,$roles) > 1) {
-                $this->perfil($id);
-            } else {
-                echo "<script>
-                        i=5;
-                            setInterval(function() {
-                                if (i==0) {
-                                    location.href='index.php';
-                                }
-                            document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
-                                i--;
-                            },1000);
-                    </script>";
-            }
+            if ($this->seguridad->esAdmin()) {
+                $id = $this->usuario->getLastId();
+                $usuario = $_REQUEST["usuario"];
+                $contrasenya = $_REQUEST["contrasenya"];
+                $email = $_REQUEST["email"];
+                $nombre = $_REQUEST["nombre"];
+                $apellido1 = $_REQUEST["apellido1"];
+                $apellido2 = $_REQUEST["apellido2"];
+                $dni = $_REQUEST["dni"];
+                $imagen = $_FILES["imagen"];
+                $borrado = 'no';
+                if (isset($_REQUEST["roles"])) {
+                    $roles = $_REQUEST["roles"];
+                } else {
+                    $roles = array("2");
+                }
 
+                if ($this->usuario->add($id,$usuario,$contrasenya,$email,$nombre,$apellido1,$apellido2,$dni,$imagen,$borrado,$roles) > 1) {
+                    $this->perfil($id);
+                } else {
+                    echo "<script>
+                            i=5;
+                                setInterval(function() {
+                                    if (i==0) {
+                                        location.href='index.php';
+                                    }
+                                document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
+                                    i--;
+                                },1000);
+                        </script>";
+                } 
+            } else {
+                $this->gestionReservas();
+            }
+            
         }
 
         public function perfil($id = null) {
 
-            if (!isset($id)) {
-                $id = $_REQUEST["id"];
+            if ($this->seguridad->esAdmin()) {
+
+                if (!isset($id)) {
+                    $id = $_REQUEST["id"];
+                }
+            
+                $data["todosLosRoles"] = $this->rol->getAll();
+                $data["rolesUsuario"] = $this->rol->get($id);
+                $data["usuario"] = $this->usuario->get($id);
+                $data["reservas"] = $this->usuario->getReservas($id);
+                $this->vista->mostrar("usuario/perfil",$data);
+
+            } else {
+                if (isset($id)) {
+                    if ($id == $this->seguridad->get('id')) {
+                        $data["todosLosRoles"] = $this->rol->getAll();
+                        $data["rolesUsuario"] = $this->rol->get($id);
+                        $data["usuario"] = $this->usuario->get($id);
+                        $data["reservas"] = $this->usuario->getReservas($id);
+                        $this->vista->mostrar("usuario/perfil",$data);
+                    }
+                    $id = $this->seguridad->get('id');
+                } else {
+                    $this->gestionReservas();
+                }
             }
-        
-            $data["todosLosRoles"] = $this->rol->getAll();
-            $data["rolesUsuario"] = $this->rol->get($id);
-            $data["usuario"] = $this->usuario->get($id);
-            $data["reservas"] = $this->usuario->getReservas($id);
-            $this->vista->mostrar("usuario/perfil",$data);
 
         }
 
         public function modificarUsuario() {
 
-            $id = $_REQUEST["id"];
-            $usuario = $_REQUEST["usuario"];
-            $contrasenya = $_REQUEST["contrasenya"];
-            $email = $_REQUEST["email"];
-            $nombre = $_REQUEST["nombre"];
-            $apellido1 = $_REQUEST["apellido1"];
-            $apellido2 = $_REQUEST["apellido2"];
-            $dni = $_REQUEST["dni"];
-            $imagen = $_FILES["imagen"];
-            if (isset($_REQUEST["roles"])) {
-                $roles = $_REQUEST["roles"];
-            } else {
-                $roles = array("2");
-            }
-            
+            if ($this->seguridad->esAdmin() || $_REQUEST["id"] == $this->seguridad->get('id')) {
+                $id = $_REQUEST["id"];
+                $usuario = $_REQUEST["usuario"];
+                $contrasenya = $_REQUEST["contrasenya"];
+                $email = $_REQUEST["email"];
+                $nombre = $_REQUEST["nombre"];
+                $apellido1 = $_REQUEST["apellido1"];
+                $apellido2 = $_REQUEST["apellido2"];
+                $dni = $_REQUEST["dni"];
+                $imagen = $_FILES["imagen"];
+                if (isset($_REQUEST["roles"])) {
+                    $roles = $_REQUEST["roles"];
+                } else {
+                    $roles = array("2");
+                }
+                
 
-            if ($this->usuario->update($id,$usuario,$contrasenya,$email,$nombre,$apellido1,$apellido2,$dni,$imagen,$roles) > 1) {
-                $this->perfil($id);
+                if ($this->usuario->update($id,$usuario,$contrasenya,$email,$nombre,$apellido1,$apellido2,$dni,$imagen,$roles) > 1) {
+                    $this->perfil($id);
+                } else {
+                    echo "<script>
+                            i=5;
+                                setInterval(function() {
+                                    if (i==0) {
+                                        location.href='index.php';
+                                    }
+                                document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
+                                    i--;
+                                },1000);
+                        </script>";
+                }
             } else {
-                echo "<script>
-                        i=5;
-                            setInterval(function() {
-                                if (i==0) {
-                                    location.href='index.php';
-                                }
-                            document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
-                                i--;
-                            },1000);
-                    </script>";
+                $this->gestionReservas();
             }
 
         }
 
         public function eliminarUsuario() {
 
-            $id = $_REQUEST["id"];
-            $this->usuario->delete($id);
-
-            $this->gestionUsuarios();
+            if ($this->seguridad->esAdmin() || $_REQUEST["id"] == $this->seguridad->get('id')) {
+                $id = $_REQUEST["id"];
+                $this->usuario->delete($id);
+                if ($id == $this->seguridad->get('id')) {
+                    $this->cerrarSesion();
+                }
+                $this->gestionUsuarios();
+            } else {
+                $this->gestionReservas();
+            }
+            
 
         }
 
@@ -297,57 +346,65 @@
 
         public function crearInstalacion() {
 
-            $id = $this->instalacion->getLastId();
-            $nombre = $_REQUEST["nombre"];
-            $descripcion = $_REQUEST["descripcion"];
-            $imagen = $_FILES["imagen"];
-            $precioHora = $_REQUEST["precioHora"];
-            $horarios = array();
-
-            for ($i=1; $i<=14; $i++) {
-                $horarios[] = $_REQUEST["d" . $i];
-            }
-
-
-            if ($this->instalacion->add($nombre,$descripcion,$imagen,$precioHora,$horarios,$id) > 1) {
-                $this->instalacion($id);
+            if ($this->seguridad->esAdmin()) {
+                $id = $this->instalacion->getLastId();
+                $nombre = $_REQUEST["nombre"];
+                $descripcion = $_REQUEST["descripcion"];
+                $imagen = $_FILES["imagen"];
+                $precioHora = $_REQUEST["precioHora"];
+                $horarios = array();
+    
+                for ($i=1; $i<=14; $i++) {
+                    $horarios[] = $_REQUEST["d" . $i];
+                }
+    
+    
+                if ($this->instalacion->add($nombre,$descripcion,$imagen,$precioHora,$horarios,$id) > 1) {
+                    $this->instalacion($id);
+                } else {
+                    echo "<script>
+                            i=5;
+                                setInterval(function() {
+                                    if (i==0) {
+                                        location.href='index.php';
+                                    }
+                                document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
+                                    i--;
+                                },1000);
+                        </script>";
+                }
             } else {
-                echo "<script>
-                        i=5;
-                            setInterval(function() {
-                                if (i==0) {
-                                    location.href='index.php';
-                                }
-                            document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
-                                i--;
-                            },1000);
-                    </script>";
+                $this->gestionReservas();
             }
 
         }
 
         public function modificarInstalacion() {
 
-            $id = $_REQUEST["id"];
-            $nombre = $_REQUEST["nombre"];
-            $descripcion = $_REQUEST["descripcion"];
-            $imagen = $_FILES["imagen"];
-            $precioHora = $_REQUEST["precioHora"];
+            if ($this->seguridad->esAdmin()) {
+                $id = $_REQUEST["id"];
+                $nombre = $_REQUEST["nombre"];
+                $descripcion = $_REQUEST["descripcion"];
+                $imagen = $_FILES["imagen"];
+                $precioHora = $_REQUEST["precioHora"];
 
 
-            if ($this->instalacion->update($id,$nombre,$descripcion,$imagen,$precioHora) > 0) {
-                $this->instalacion($id);
+                if ($this->instalacion->update($id,$nombre,$descripcion,$imagen,$precioHora) > 0) {
+                    $this->instalacion($id);
+                } else {
+                    echo "<script>
+                            i=5;
+                                setInterval(function() {
+                                    if (i==0) {
+                                        location.href='index.php?action=gestionInstalaciones';
+                                    }
+                                document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
+                                    i--;
+                                },1000);
+                        </script>";
+                }
             } else {
-                echo "<script>
-                        i=5;
-                            setInterval(function() {
-                                if (i==0) {
-                                    location.href='index.php?action=gestionInstalaciones';
-                                }
-                            document.body.innerHTML = 'Ha ocurrido un error. Redireccionando en ' + i;
-                                i--;
-                            },1000);
-                    </script>";
+                $this->gestionReservas();
             }
 
         }
@@ -373,22 +430,31 @@
 
         public function instalacion($id = null) {
 
-            if (!isset($id)) {
-                $id = $_REQUEST["id"];
+            if ($this->seguridad->esAdmin()) {
+                if (!isset($id)) {
+                    $id = $_REQUEST["id"];
+                }
+    
+                $data["instalacion"] = $this->instalacion->get($id);
+                $data["horarios"] = $this->horario->get($id);
+                $this->vista->mostrar("instalacion/perfilInstalacion",$data);
+            } else {
+                $this->gestionReservas();
             }
-
-            $data["instalacion"] = $this->instalacion->get($id);
-            $data["horarios"] = $this->horario->get($id);
-            $this->vista->mostrar("instalacion/perfilInstalacion",$data);
 
         }
 
         public function eliminarInstalacion() {
 
-            $id = $_REQUEST["id"];
-            $this->instalacion->delete($id);
-
-            $this->gestionInstalaciones();
+            if ($this->seguridad->esAdmin()) {
+                $id = $_REQUEST["id"];
+                $this->instalacion->delete($id);
+    
+                $this->gestionInstalaciones();
+            } else {
+                $this->gestionReservas();
+            }
+            
 
         }
 
@@ -398,11 +464,12 @@
 
             $id = $this->reserva->getLastId();
             $fecha = $_REQUEST["fecha"];
-            $hora = $_REQUEST["hora"];
+            $hora = '00:00';
             $instalacion = $_REQUEST["instalacion"];
             $precio = $this->instalacion->getCampo($instalacion,'precioHora');
+            $usuario = $_REQUEST["idUsuario"];
 
-            if ($this->reserva->add($id, $fecha, $hora, $precio, $instalacion) > 0) {
+            if ($this->reserva->add($id, $fecha, $hora, $precio, $instalacion, $usuario) > 0) {
                 $this->gestionReservas();
             } else {
                 echo "<script>
